Fix issue #4036 - DbContext.X.Where(x => x.YRef.ID == 1) throws "The binary operator Equal is not defined for the types Nullable..."
Issue was that the navigation property -> foreign key optimization was too aggressive. Problem with the optimization is that sometimes FK is of different type than PK, and it may cause type inconsistencies.
We incorrectly assumed that it can be performed for required navigations. However one can declare required navigation that has nullable FK, while PK is non-nullable.

Fix is to compare types of PK and FK instead.